feat: move the config.php configuration file for the database into a config subdirectory folder
