+++ database.php
@@ -9,7 +9,7 @@ class Database {
     // Constructor initializes and creates the connection to the database
     public function __construct($host = null, $dbName = null, $username = null, $password = null) {
         // Include the configuration file with the database credentials
-        require_once "config.php";
+        require_once "config/config.php";
 
         // Use provided parameters or fallback to constants from the configuration file
         $this->host = $host ?? DB_HOSTNAME;
